+++ hutool-core/src/main/java/com/xiaoleilu/hutool/convert/Convert.java
@@ -8,6 +8,7 @@ import java.util.Set;
 import java.util.concurrent.TimeUnit;
 
 import com.xiaoleilu.hutool.convert.impl.CollectionConverter;
+import com.xiaoleilu.hutool.convert.impl.EnumConverter;
 import com.xiaoleilu.hutool.lang.Assert;
 import com.xiaoleilu.hutool.util.CharsetUtil;
 import com.xiaoleilu.hutool.util.HexUtil;
@@ -45,6 +46,17 @@ public final class Convert {
 	public static String toStr(Object value) {
 		return toStr(value, null);
 	}
+	
+	/**
+	 * 转换为String数组
+	 * 
+	 * @param value 被转换的值
+	 * @return String数组
+	 * @since 3.2.0
+	 */
+	public static String[] toStrArray(Object value) {
+		return convert(String[].class, value);
+	}
 
 	/**
 	 * 转换为字符<br>
@@ -69,6 +81,17 @@ public final class Convert {
 	 */
 	public static Character toChar(Object value) {
 		return toChar(value, null);
+	}
+	
+	/**
+	 * 转换为Character数组
+	 * 
+	 * @param value 被转换的值
+	 * @return Character数组
+	 * @since 3.2.0
+	 */
+	public static Character[] toCharArray(Object value) {
+		return convert(Character[].class, value);
 	}
 
 	/**
@@ -95,6 +118,17 @@ public final class Convert {
 	public static Byte toByte(Object value) {
 		return toByte(value, null);
 	}
+	
+	/**
+	 * 转换为Byte数组
+	 * 
+	 * @param value 被转换的值
+	 * @return Byte数组
+	 * @since 3.2.0
+	 */
+	public static Byte[] toByteArray(Object value) {
+		return convert(Byte[].class, value);
+	}
 
 	/**
 	 * 转换为Short<br>
@@ -120,6 +154,17 @@ public final class Convert {
 	public static Short toShort(Object value) {
 		return toShort(value, null);
 	}
+	
+	/**
+	 * 转换为Short数组
+	 * 
+	 * @param value 被转换的值
+	 * @return Short数组
+	 * @since 3.2.0
+	 */
+	public static Short[] toShortArray(Object value) {
+		return convert(Short[].class, value);
+	}
 
 	/**
 	 * 转换为Number<br>
@@ -144,6 +189,17 @@ public final class Convert {
 	 */
 	public static Number toNumber(Object value) {
 		return toNumber(value, null);
+	}
+	
+	/**
+	 * 转换为Number数组
+	 * 
+	 * @param value 被转换的值
+	 * @return Number数组
+	 * @since 3.2.0
+	 */
+	public static Number[] toNumberArray(Object value) {
+		return convert(Number[].class, value);
 	}
 
 	/**
@@ -382,23 +438,7 @@ public final class Convert {
 	 * @return Enum
 	 */
 	public static <E extends Enum<E>> E toEnum(Class<E> clazz, Object value, E defaultValue) {
-		if (value == null) {
-			return defaultValue;
-		}
-		if (clazz.isAssignableFrom(value.getClass())) {
-			@SuppressWarnings("unchecked")
-			E myE = (E) value;
-			return myE;
-		}
-		final String valueStr = toStr(value, null);
-		if (StrUtil.isBlank(valueStr)) {
-			return defaultValue;
-		}
-		try {
-			return Enum.valueOf(clazz, valueStr);
-		} catch (Exception e) {
-			return defaultValue;
-		}
+		return (new EnumConverter<>(clazz)).convert(value, defaultValue);
 	}
 
 	/**
@@ -413,16 +453,17 @@ public final class Convert {
 	public static <E extends Enum<E>> E toEnum(Class<E> clazz, Object value) {
 		return toEnum(clazz, value, null);
 	}
-	
+
 	/**
 	 * 转换为集合类
+	 * 
 	 * @param collectionType 集合类型
 	 * @param elementType 集合中元素类型
 	 * @param value 被转换的值
 	 * @return {@link Collection}
 	 * @since 3.0.8
 	 */
-	public static Collection<?> toCollection(Class<?> collectionType, Class<?> elementType, Object value){
+	public static Collection<?> toCollection(Class<?> collectionType, Class<?> elementType, Object value) {
 		return new CollectionConverter(collectionType, elementType).convert(value, null);
 	}
 
@@ -514,7 +555,7 @@ public final class Convert {
 			}
 
 			if (c[i] == '\u3000' || c[i] == '\u00a0' || c[i] == '\u2007' || c[i] == '\u202F') {
-				//\u3000是中文全角空格，\u00a0、\u2007、\u202F是不间断空格
+				// \u3000是中文全角空格，\u00a0、\u2007、\u202F是不间断空格
 				c[i] = ' ';
 			} else if (c[i] > '\uFF00' && c[i] < '\uFF5F') {
 				c[i] = (char) (c[i] - 65248);
@@ -587,8 +628,8 @@ public final class Convert {
 			strHex = Integer.toHexString(strText.charAt(i));
 			strHexLen = strHex.length();
 			str.append("\\u");
-			//对不够4位的在前补零
-			if(strHexLen > 0 && strHexLen < 4){
+			// 对不够4位的在前补零
+			if (strHexLen > 0 && strHexLen < 4) {
 				str.append(StrUtil.repeat('0', 4 - strHexLen));
 			}
 			str.append(strHex);
@@ -702,15 +743,130 @@ public final class Convert {
 	public static Class<?> unWrap(Class<?> clazz) {
 		return BasicType.unWrap(clazz);
 	}
-	
-	//-------------------------------------------------------------------------- 数字和英文转换
+
+	// -------------------------------------------------------------------------- 数字和英文转换
 	/**
 	 * 将阿拉伯数字转为英文表达式
+	 * 
 	 * @param number {@link Number}对象
 	 * @return 英文表达式
 	 * @since 3.0.9
 	 */
 	public static String numberToWord(Number number) {
 		return NumberWordFormater.format(number);
+	}
+
+	// -------------------------------------------------------------------------- 数字转换
+	/**
+	 * int转byte
+	 * 
+	 * @param intValue int值
+	 * @return byte值
+	 * @since 3.2.0
+	 */
+	public static byte intToByte(int intValue) {
+		return (byte) intValue;
+	}
+
+	/**
+	 * byte转无符号int
+	 * 
+	 * @param byteValue byte值
+	 * @return 无符号int值
+	 * @since 3.2.0
+	 */
+	public static int byteToUnsignedInt(byte byteValue) {
+		// Java 总是把 byte 当做有符处理；我们可以通过将其和 0xFF 进行二进制与得到它的无符值
+		return byteValue & 0xFF;
+	}
+
+	/**
+	 * byte数组转short
+	 * 
+	 * @param bytes byte数组
+	 * @return short值
+	 * @since 3.2.0
+	 */
+	public static short bytesToShort(byte[] bytes) {
+		return (short) (bytes[1] & 0xff | (bytes[0] & 0xff) << 8);
+	}
+
+	/**
+	 * short转byte数组
+	 * @param shortValue short值
+	 * @return byte数组
+	 * @since 3.2.0
+	 */
+	public static byte[] shortToBytes(short shortValue) {
+		byte[] b = new byte[2];
+		b[1] = (byte) (shortValue & 0xff);
+		b[0] = (byte) ((shortValue >> 8) & 0xff);
+		return b;
+	}
+
+	/**
+	 * byte[]转int值
+	 * 
+	 * @param bytes byte数组
+	 * @return int值
+	 * @since 3.2.0
+	 */
+	public static int bytesToInt(byte[] bytes) {
+		return bytes[3] & 0xFF | //
+				(bytes[2] & 0xFF) << 8 | //
+				(bytes[1] & 0xFF) << 16 | //
+				(bytes[0] & 0xFF) << 24; //
+	}
+
+	/**
+	 * int转byte数组
+	 * 
+	 * @param intValue int值
+	 * @return byte数组
+	 * @since 3.2.0
+	 */
+	public static byte[] intToBytes(int intValue) {
+		return new byte[] { //
+				(byte) ((intValue >> 24) & 0xFF), //
+				(byte) ((intValue >> 16) & 0xFF), //
+				(byte) ((intValue >> 8) & 0xFF), //
+				(byte) (intValue & 0xFF) //
+		};
+	}
+
+	/**
+	 * long转byte数组<br>
+	 * from: https://stackoverflow.com/questions/4485128/how-do-i-convert-long-to-byte-and-back-in-java
+	 * 
+	 * @param longValue long值
+	 * @return byte数组
+	 * @since 3.2.0
+	 */
+	public static byte[] longToBytes(long longValue) {
+		// Magic number 8 should be defined as Long.SIZE / Byte.SIZE
+		final byte[] result = new byte[8];
+		for (int i = 7; i >= 0; i--) {
+			result[i] = (byte) (longValue & 0xFF);
+			longValue >>= 8;
+		}
+		return result;
+	}
+
+	/**
+	 * byte数组转long<br>
+	 * from: https://stackoverflow.com/questions/4485128/how-do-i-convert-long-to-byte-and-back-in-java
+	 * 
+	 * @param bytes byte数组
+	 * @return long值
+	 * @since 3.2.0
+	 */
+	public static long bytesToLong(byte[] bytes) {
+		// Magic number 8 should be defined as Long.SIZE / Byte.SIZE
+		long values = 0;
+		for (int i = 0; i < 8; i++) {
+			values <<= 8;
+			values |= (bytes[i] & 0xff);
+		}
+		return values;
 	}
 }