+++ hutool-core/src/main/java/com/xiaoleilu/hutool/util/StrUtil.java
@@ -555,6 +555,38 @@ public class StrUtil {
 	public static boolean contains(CharSequence str, char searchChar) {
 		return indexOf(str, searchChar) > -1;
 	}
+	
+	/**
+	 * 查找指定字符串是否包含指定字符串列表中的任意一个字符串
+	 * 
+	 * @param str 指定字符串
+	 * @param testStrs 需要检查的字符串数组
+	 * @return 是否包含任意一个字符串
+	 * @since 3.2.0
+	 */
+	public static boolean containsAny(CharSequence str, CharSequence... testStrs) {
+		return null != getContainsStr(str, testStrs);
+	}
+	
+	/**
+	 * 查找指定字符串是否包含指定字符串列表中的任意一个字符串，如果包含返回找到的第一个字符串
+	 * 
+	 * @param str 指定字符串
+	 * @param testStrs 需要检查的字符串数组
+	 * @return 被包含的第一个字符串
+	 * @since 3.2.0
+	 */
+	public static String getContainsStr(CharSequence str, CharSequence... testStrs) {
+		if(isEmpty(str) || ArrayUtil.isEmpty(testStrs)) {
+			return null;
+		}
+		for (CharSequence checkStr : testStrs) {
+			if(str.toString().contains(checkStr)) {
+				return checkStr.toString();
+			}
+		}
+		return null;
+	}
 
 	/**
 	 * 是否包含特定字符，忽略大小写，如果给定两个参数都为<code>null</code>，返回true
@@ -569,6 +601,40 @@ public class StrUtil {
 			return null == testStr;
 		}
 		return str.toString().toLowerCase().contains(testStr.toString().toLowerCase());
+	}
+	
+	/**
+	 * 查找指定字符串是否包含指定字符串列表中的任意一个字符串<br>
+	 * 忽略大小写
+	 * 
+	 * @param str 指定字符串
+	 * @param testStrs 需要检查的字符串数组
+	 * @return 是否包含任意一个字符串
+	 * @since 3.2.0
+	 */
+	public static boolean containsAnyIgnoreCase(CharSequence str, CharSequence... testStrs) {
+		return null != getContainsStrIgnoreCase(str, testStrs);
+	}
+	
+	/**
+	 * 查找指定字符串是否包含指定字符串列表中的任意一个字符串，如果包含返回找到的第一个字符串<br>
+	 * 忽略大小写
+	 * 
+	 * @param str 指定字符串
+	 * @param testStrs 需要检查的字符串数组
+	 * @return 被包含的第一个字符串
+	 * @since 3.2.0
+	 */
+	public static String getContainsStrIgnoreCase(CharSequence str, CharSequence... testStrs) {
+		if(isEmpty(str) || ArrayUtil.isEmpty(testStrs)) {
+			return null;
+		}
+		for (CharSequence testStr : testStrs) {
+			if(containsIgnoreCase(str, testStr)) {
+				return testStr.toString();
+			}
+		}
+		return null;
 	}
 
 	/**
@@ -933,6 +999,7 @@ public class StrUtil {
 		return sb.toString();
 	}
 
+	//------------------------------------------------------------------------------ Split
 	/**
 	 * 切分字符串
 	 * 
@@ -995,6 +1062,18 @@ public class StrUtil {
 	public static List<String> splitTrim(CharSequence str, char separator) {
 		return splitTrim(str, separator, -1);
 	}
+	
+	/**
+	 * 切分字符串，去除切分后每个元素两边的空白符，去除空白项
+	 * 
+	 * @param str 被切分的字符串
+	 * @param separator 分隔符字符
+	 * @return 切分后的集合
+	 * @since 3.2.0
+	 */
+	public static List<String> splitTrim(CharSequence str, CharSequence separator) {
+		return splitTrim(str, separator, -1);
+	}
 
 	/**
 	 * 切分字符串，去除切分后每个元素两边的空白符，去除空白项
@@ -1006,6 +1085,19 @@ public class StrUtil {
 	 * @since 3.1.0
 	 */
 	public static List<String> splitTrim(CharSequence str, char separator, int limit) {
+		return split(str, separator, limit, true, true);
+	}
+	
+	/**
+	 * 切分字符串，去除切分后每个元素两边的空白符，去除空白项
+	 * 
+	 * @param str 被切分的字符串
+	 * @param separator 分隔符字符
+	 * @param limit 限制分片数，-1不限制
+	 * @return 切分后的集合
+	 * @since 3.2.0
+	 */
+	public static List<String> splitTrim(CharSequence str, CharSequence separator, int limit) {
 		return split(str, separator, limit, true, true);
 	}
 
@@ -1039,6 +1131,25 @@ public class StrUtil {
 			return new ArrayList<>(0);
 		}
 		return StrSpliter.split(str.toString(), separator, limit, isTrim, ignoreEmpty);
+	}
+	
+	/**
+	 * 切分字符串
+	 * 
+	 * @param str 被切分的字符串
+	 * @param separator 分隔符字符
+	 * @param limit 限制分片数，-1不限制
+	 * @param isTrim 是否去除切分字符串后每个元素两边的空格
+	 * @param ignoreEmpty 是否忽略空串
+	 * @return 切分后的集合
+	 * @since 3.2.0
+	 */
+	public static List<String> split(CharSequence str, CharSequence separator, int limit, boolean isTrim, boolean ignoreEmpty) {
+		if (null == str) {
+			return new ArrayList<>(0);
+		}
+		final String separatorStr = (null == separator) ? null : separator.toString();
+		return StrSpliter.split(str.toString(), separatorStr, limit, isTrim, ignoreEmpty);
 	}
 
 	/**
@@ -1434,11 +1545,7 @@ public class StrUtil {
 	 * @return 如果两个字符串相同，或者都是<code>null</code>，则返回<code>true</code>
 	 */
 	public static boolean equals(CharSequence str1, CharSequence str2) {
-		if (str1 == null) {
-			return str2 == null;
-		}
-
-		return str1.equals(str2);
+		return equals(str1, str2, false);
 	}
 
 	/**
@@ -1458,11 +1565,28 @@ public class StrUtil {
 	 * @return 如果两个字符串相同，或者都是<code>null</code>，则返回<code>true</code>
 	 */
 	public static boolean equalsIgnoreCase(CharSequence str1, CharSequence str2) {
+		return equals(str1, str2, true);
+	}
+	
+	/**
+	 * 比较两个字符串是否相等。
+	 * 
+	 * @param str1 要比较的字符串1
+	 * @param str2 要比较的字符串2
+	 * @param ignoreCase 是否忽略大小写
+	 * @return 如果两个字符串相同，或者都是<code>null</code>，则返回<code>true</code>
+	 * @since 3.2.0
+	 */
+	public static boolean equals(CharSequence str1, CharSequence str2, boolean ignoreCase) {
 		if (str1 == null) {
 			return str2 == null;
 		}
-
-		return str1.toString().equalsIgnoreCase(str2.toString());
+		
+		if(ignoreCase) {
+			return str1.toString().equalsIgnoreCase(str2.toString());
+		}else {
+			return str1.equals(str2);
+		}
 	}
 
 	/**