+++ hutool-poi/src/main/java/com/xiaoleilu/hutool/poi/excel/ExcelReader.java
@@ -1,5 +1,6 @@
 package com.xiaoleilu.hutool.poi.excel;
 
+import java.io.Closeable;
 import java.io.File;
 import java.io.InputStream;
 import java.util.ArrayList;
@@ -11,9 +12,11 @@ import org.apache.poi.ss.usermodel.Row;
 import org.apache.poi.ss.usermodel.Sheet;
 import org.apache.poi.ss.usermodel.Workbook;
 
-import com.xiaoleilu.hutool.collection.IterUtil;
-import com.xiaoleilu.hutool.poi.excel.editors.TrimEditor;
 import com.xiaoleilu.hutool.bean.BeanUtil;
+import com.xiaoleilu.hutool.collection.IterUtil;
+import com.xiaoleilu.hutool.io.IoUtil;
+import com.xiaoleilu.hutool.lang.Assert;
+import com.xiaoleilu.hutool.poi.excel.editors.TrimEditor;
 import com.xiaoleilu.hutool.util.StrUtil;
 
 /**
@@ -23,8 +26,12 @@ import com.xiaoleilu.hutool.util.StrUtil;
  * @author Looly
  * @since 3.1.0
  */
-public class ExcelReader {
+public class ExcelReader implements Closeable{
 
+	/** 是否被关闭 */
+	private boolean isClosed;
+	/** 工作簿 */
+	private Workbook workbook;
 	/** Excel中对应的Sheet */
 	private Sheet sheet;
 
@@ -104,6 +111,7 @@ public class ExcelReader {
 	 */
 	public ExcelReader(Workbook book, String sheetName) {
 		this(book.getSheet(sheetName));
+		this.workbook = book;
 	}
 
 	/**
@@ -211,6 +219,7 @@ public class ExcelReader {
 	 * @return 行的集合，一行使用List表示
 	 */
 	public List<List<Object>> read(int startRowIndex, int endRowIndex) {
+		Assert.isFalse(this.isClosed, "ExcelReader has been closed!");
 		List<List<Object>> resultList = new ArrayList<>();
 
 		startRowIndex = Math.max(startRowIndex, sheet.getFirstRowNum());// 读取起始行（包含）
@@ -245,6 +254,7 @@ public class ExcelReader {
 	 * @return Map的列表
 	 */
 	public List<Map<String, Object>> read(int headerRowIndex, int startRowIndex, int endRowIndex) {
+		Assert.isFalse(this.isClosed, "ExcelReader has been closed!");
 		// 边界判断
 		final int firstRowNum = sheet.getFirstRowNum();
 		final int lastRowNum = sheet.getLastRowNum();
@@ -296,6 +306,7 @@ public class ExcelReader {
 	 */
 	@SuppressWarnings("unchecked")
 	public <T> List<T> read(int headerRowIndex, int startRowIndex, int endRowIndex, Class<T> beanType) {
+		Assert.isFalse(this.isClosed, "ExcelReader has been closed!");
 		final List<Map<String, Object>> mapList = read(headerRowIndex, startRowIndex, endRowIndex);
 		if (Map.class.isAssignableFrom(beanType)) {
 			return (List<T>) mapList;
@@ -307,6 +318,18 @@ public class ExcelReader {
 		}
 		return beanList;
 	}
+	
+	/**
+	 * 关闭工作簿
+	 * @since 3.2.0
+	 */
+	@Override
+	public void close() {
+		IoUtil.close(this.workbook);
+		this.sheet = null;
+		this.workbook = null;
+		this.isClosed = true;
+	}
 
 	// ------------------------------------------------------------------------------------------------------- Private methods start
 	/**
@@ -316,13 +339,7 @@ public class ExcelReader {
 	 * @return 单元格值列表
 	 */
 	private List<Object> readRow(Row row) {
-		final List<Object> cellValues = new ArrayList<>();
-		
-		short length = row.getLastCellNum();
-		for (short i = 0; i < length; i++) {
-			cellValues.add(ExcelUtil.getCellValue(row.getCell(i), cellEditor));
-		}
-		return cellValues;
+		return InternalExcelUtil.readRow(row, this.cellEditor);
 	}
 
 	/**