+++ hutool-poi/src/main/java/com/xiaoleilu/hutool/poi/excel/sax/handler/RowHandler.java
@@ -1,4 +1,4 @@
-package com.xiaoleilu.hutool.poi.excel.sax;
+package com.xiaoleilu.hutool.poi.excel.sax.handler;
 
 import java.util.List;
 
@@ -13,7 +13,7 @@ public interface RowHandler {
 	 * 处理一行数据
 	 * @param sheetIndex 当前Sheet序号
 	 * @param rowIndex 当前行号
-	 * @param rowlist 行数据列表
+	 * @param rowList 行数据列表
 	 */
-	void handle(int sheetIndex, int rowIndex, List<String> rowlist);
+	void handle(int sheetIndex, int rowIndex, List<Object> rowList);
 }